requirements.txt
numpy
scikit-learn
scikit-image
pillow
tensorflow
keras
pynput
h5py
